+++ demo.rb
@@ -26,8 +26,8 @@ puts '[1, 2, 3, 4, 5].my_none? { |i| i % 11 == 0 }'
 p(a.my_none? { |i| (i % 11).zero? })
 puts '============'
 
-puts '[1, 2, 3, 4, 5].my_count?(99)'
-p(a.my_count?(99))
+puts '[1, 2, 3, 4, 5].my_count(99)'
+p(a.my_count(99))
 puts '============'
 
 puts '[1, 2, 3, 4, 5].my_map? { |item| item * 2 }'
@@ -39,8 +39,8 @@ p(a.my_map! { |item| item + 10 })
 a.my_map! { |item| item - 10 }
 puts '============'
 
-puts '[1, 2, 3, 4, 5].inject'
-p(a.inject(:+))
+puts '[1, 2, 3, 4, 5].my_inject'
+p(a.my_inject(:+))
 puts '============'
 
 puts 'multiply_els([2, 4, 5])'